Move workspace functions to cli-xtask
