+++ examples/bin-with-command/xtask/src/main.rs
@@ -6,7 +6,7 @@ fn main() -> eyre::Result<()> {
     cli_xtask::install_error_handler()?;
     cli_xtask::install_logger()?;
 
-    let metadata = cli_xtask::cargo_workspace();
+    let metadata = cli_xtask::workspace::current();
     let (dist, package) = DistConfigBuilder::from_root_package(metadata)?;
     let dist = dist
         .package(package.binary_from_command(Args::command())?.build())
