+++ xtask/Cargo.toml
@@ -13,4 +13,3 @@ color-eyre = "0.6.2"
 eyre = "0.6.8"
 tracing = "0.1.36"
 tracing-subscriber = { version = "0.3.15", features = ["env-filter"] }
-walkdir = "2.3.2"
