+++ xtask/src/clippy.rs
@@ -1,3 +1,5 @@
+use cli_xtask::workspace;
+
 #[derive(Debug, clap::Parser)]
 pub(crate) struct Args {
     /// Arguments to pass to the `cargo clippy`
@@ -9,11 +11,11 @@ impl Args {
     pub(crate) fn run(&self) -> eyre::Result<()> {
         let Self { extra_options } = self;
 
-        for (_path, metadata) in crate::all_workspaces()? {
+        for metadata in workspace::all() {
             for package in metadata.workspace_packages() {
                 for feature_args in crate::feature_combinations(package) {
                     crate::execute_on(
-                        &metadata,
+                        metadata,
                         "cargo",
                         ["clippy", "--all-targets", "--package", &package.name]
                             .into_iter()
