+++ xtask/src/rdme.rs
@@ -1,3 +1,5 @@
+use cli_xtask::workspace;
+
 #[derive(Debug, clap::Parser)]
 pub(crate) struct Args {
     /// Arguments to pass to the `cargo rdme`
@@ -9,9 +11,9 @@ impl Args {
     pub(crate) fn run(&self) -> eyre::Result<()> {
         let Self { extra_options } = self;
 
-        for (_path, metadata) in crate::all_workspaces()? {
+        for metadata in workspace::all() {
             crate::execute_on(
-                &metadata,
+                metadata,
                 "cargo",
                 ["rdme"]
                     .into_iter()
