+++ xtask/src/test.rs
@@ -1,3 +1,5 @@
+use cli_xtask::workspace;
+
 #[derive(Debug, clap::Parser)]
 pub(crate) struct Args {
     /// Arguments to pass to the `cargo test`
@@ -9,13 +11,13 @@ impl Args {
     pub(crate) fn run(&self) -> eyre::Result<()> {
         let Self { extra_options } = self;
 
-        for (_path, metadata) in crate::all_workspaces()? {
+        for metadata in workspace::all() {
             for package in metadata.workspace_packages() {
                 for feature_args in crate::feature_combinations(package) {
                     // cargo test --package <pkg> <features> <extra_options>
                     // DO NOT USE `--all-targets` here, doctests are not built with `--all-targets`
                     crate::execute_on(
-                        &metadata,
+                        metadata,
                         "cargo",
                         ["test", "--package", &package.name]
                             .into_iter()
