+++ xtask/src/udeps.rs
@@ -1,3 +1,5 @@
+use cli_xtask::workspace;
+
 #[derive(Debug, clap::Parser)]
 pub(crate) struct Args {
     /// Arguments to pass to the `cargo udeps`
@@ -9,12 +11,12 @@ impl Args {
     pub(crate) fn run(&self) -> eyre::Result<()> {
         let Self { extra_options } = self;
 
-        for (_path, metadata) in crate::all_workspaces()? {
+        for metadata in workspace::all() {
             for package in metadata.workspace_packages() {
                 for feature_args in crate::feature_combinations(package) {
                     // cargo +nightly udeps fails on windows, so use rustup instead
                     crate::execute_on(
-                        &metadata,
+                        metadata,
                         "rustup",
                         [
                             "run",
